+++ db/sqlc/account.sql.go
@@ -51,7 +51,7 @@ const deleteAccount = `-- name: DeleteAccount :exec
 DELETE FROM accounts WHERE id = $1
 `
 
-func (q *Queries) DeleteAccount(ctx context.Context, id int32) error {
+func (q *Queries) DeleteAccount(ctx context.Context, id int64) error {
 	_, err := q.db.ExecContext(ctx, deleteAccount, id)
 	return err
 }
@@ -60,7 +60,7 @@ const getAccount = `-- name: GetAccount :one
 SELECT id, username, type, document, email, password, balance, created_at FROM accounts WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetAccount(ctx context.Context, id int32) (Account, error) {
+func (q *Queries) GetAccount(ctx context.Context, id int64) (Account, error) {
 	row := q.db.QueryRowContext(ctx, getAccount, id)
 	var i Account
 	err := row.Scan(
@@ -120,8 +120,8 @@ RETURNING id, username, type, document, email, password, balance, created_at
 `
 
 type UpdateAccountParams struct {
-	ID      int32 `json:"id"`
-	Balance int32 `json:"balance"`
+	ID      int64 `json:"id"`
+	Balance int64 `json:"balance"`
 }
 
 func (q *Queries) UpdateAccount(ctx context.Context, arg UpdateAccountParams) (Account, error) {
